Final review comment changes
