+++ src/app/components/audio-recordings/pages/list/list.component.html
@@ -35,7 +35,6 @@
     <ngx-datatable-column name="Uploader">
       <ng-template let-value="value" ngx-datatable-cell-template>
         <baw-user-link [user]="value.uploader"></baw-user-link>
-        <baw-ghost-user-hint [user]="value.uploader"></baw-ghost-user-hint>
       </ng-template>
     </ngx-datatable-column>
     <ngx-datatable-column name="Site">
