+++ src/app/components/shared/user-link/user-link/user-link.component.ts
@@ -1,5 +1,7 @@
 import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
 import { User } from "@models/User";
+import { Placement } from "@ng-bootstrap/ng-bootstrap";
+import { IconProp } from "@fortawesome/fontawesome-svg-core";
 
 @Component({
   selector: "baw-user-link",
@@ -20,12 +22,28 @@ import { User } from "@models/User";
       <!-- Show ghost user -->
       <ng-template #ghostUser>
         <span [innerText]="user.userName"></span>
+
+        <fa-icon
+          *ngIf="user?.isGhost"
+          class="ms-1"
+          [icon]="icon"
+          [placement]="tooltipPlacement"
+          [ngbTooltip]="getHint()"
+        ></fa-icon>
       </ng-template>
     </ng-template>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class UserLinkComponent {
-  @Input() public user: User;
   // TODO Potentially add the ability for different styles, ie. link/badge/card
+  @Input() public user: User;
+  @Input() public tooltipPlacement: Placement = "left";
+  public icon: IconProp = ["fas", "info-circle"];
+
+  public getHint() {
+    return this.user.isUnknown
+      ? "You may not have access to this information, try logging in"
+      : "This user appears to have deleted their account";
+  }
 }
